pyproject.toml
[tool.poetry]
name = "thesis_project"
version = "0.1.0"
description = ""
authors = ["SaPer663 <saper663ru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2"
psycopg2-binary = "2.8.6"
Pillow = "^9.1.1"
djoser = "^2.1.0"
djangorestframework = "^3.13.1"
pytils = "^0.4.1"
drf-yasg = "^1.20.0"
django-filter = "^21.1"
reportlab = "^3.6.10"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
ipython = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
